story buttons variants
