+++ src/ui/button/button.module.scss
@@ -326,7 +326,7 @@
             color: calc-color(fade, $color-base, 30);
 
             .kui-button__icon {
-                color: calc-color(fade, $color-base, 30);
+                color: calc-color(fade, $color-base, 15);
             }
         }
     }
