+++ src/ui/stories/buttons/action.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import { storiesOf } from '@storybook/react';
+import { Button, ButtonsGroup } from './../../../ui';
+
+const Story = () => {    
+    return (
+        <div className="page">
+             <section>
+                <h2>Actions</h2>
+                <ButtonsGroup>
+                    <Button variant="action">Action</Button>
+                    <Button variant="action" disabled>Disabled</Button>
+                </ButtonsGroup>
+            </section>
+        </div>
+    );
+};
+
+storiesOf('Buttons', module)
+    .add('Action', () => <Story/>);
