+++ src/ui/stories/index.js
@@ -1,10 +1,21 @@
 import './stories.module.scss';
-import Icons from './icons';
-import Buttons from './buttons';
 import Snackbars from './snackbars';
 import Controls from './controls';
 
+import Action from './buttons/action';
+import Add from './buttons/add';
+import Fab from './buttons/fab';
+import Icon from './buttons/icon';
+import IconText from './buttons/iconText';
+import Primary from './buttons/primary';
+import Secondary from './buttons/secondary';
+import Text from './buttons/text';
+
+import ButtonsGroup from './buttons/buttonsGroup';
 import ButtonsSegmented from './buttons/buttonsSegmented';
+
+import Small from './icons/small';
+import Large from './icons/large';
 
 import ButtonDropdown from './controls/buttonDropdown';
 import Datepicker from './controls/datepicker';
